Add featured API and entries pagination
